requirements.txt
# OpenAI & Agents
openai-agents==0.6.2
openai

# Tipler / modeller
pydantic>=2.7

# Konfig / ortam degiskenleri
python-dotenv

# Firebase
firebase-admin>=6.5

# Google AI (Gemini API) - REST client
httpx>=0.27

# REST API
fastapi>=0.110
uvicorn>=0.30
docker

# Web scraping / search
beautifulsoup4>=4.12

# Token counting (for dry-run mode cost analysis)
tiktoken>=0.5

# fal.ai (MMAudio - video audio generation)
fal-client>=0.13

# Kling AI (JWT authentication for video generation API)
PyJWT>=2.8

# Testing
pytest>=8.0
pytest-asyncio>=0.23
